% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_discretas.R
\name{bar_discretas}
\alias{bar_discretas}
\title{Gera varios graficos}
\usage{
bar_discretas(df, grid = TRUE, axis = FALSE, fill = FALSE,
  coord_flip = FALSE, bar_fill = "gray40", bar_alpha = 1,
  bar_position = "dodge")
}
\arguments{
\item{df}{}

\item{grid}{= TRUE}

\item{axis}{= FALSE}

\item{fill}{= FALSE}

\item{coord_flip}{= FALSE}

\item{bar_fill}{= "gray40"}

\item{bar_alpha}{= 1}

\item{bar_position}{= "dodge"}
}
\value{
plot
}
\description{
Retorna uma lista de graficos
}
\examples{
conjunta_bar_cd(iris)

}
