% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_continuas.R
\name{boxplot_continuas}
\alias{boxplot_continuas}
\title{Gera varios graficos}
\usage{
boxplot_continuas(df, grid = TRUE, axis = FALSE, coord_flip = FALSE,
  col = "black", fill = "gray40", position = "dodge2",
  outlier.color = NULL, outlier.fill = NULL, outlier.shape = 19,
  outlier.size = 1.5, notch = FALSE, notchwidth = 0.5)
}
\arguments{
\item{df}{}

\item{grid}{= TRUE}

\item{axis}{= FALSE}

\item{coord_flip}{= FALSE}

\item{col}{= "black"}

\item{fill}{= "gray40"}

\item{position}{= "dodge2"}

\item{outlier.color}{= NULL}

\item{outlier.fill}{= NULL}

\item{outlier.shape}{= 19}

\item{outlier.size}{= 1.5}

\item{notch}{= FALSE}

\item{notchwidth}{= 0.5}
}
\value{
plot
}
\description{
Retorna uma lista de graficos
}
\examples{
conjunta_bar_cd(iris)

}
