% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunta_bar_cd.R
\name{conjunta_bar_cd}
\alias{conjunta_bar_cd}
\title{Gera varios graficos de barra com variaveis de um dataframe}
\usage{
conjunta_bar_cd(df, grid = TRUE, axis = FALSE, fill = FALSE,
  coord_flip = FALSE, bar_fill = "gray30", bar_alpha = 1,
  bar_position = "stack")
}
\arguments{
\item{df}{}

\item{grid}{= TRUE}

\item{axis}{= FALSE}

\item{fill}{= FALSE}

\item{coord_flip}{= FALSE}

\item{bar_fill}{= "gray30"}

\item{bar_alpha}{= 1}

\item{bar_position}{= "stack"}
}
\value{
plot
}
\description{
Retorna uma lista de graficos de barra, ou um grid de graficos de barra com as conjuntas de variaveis continuas e discretas de um dado dataframe
}
\examples{
conjunta_bar_cd(iris)

}
