% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunta_cd.R
\name{conjunta_cd}
\alias{conjunta_cd}
\title{Barras Bivariada}
\usage{
conjunta_cd(df, grid = TRUE, plot_type = "bar", coord_flip = FALSE,
  axis = FALSE, aes_fill = FALSE, col = "black", fill = "gray30",
  alpha = 1, bar_position = "stack", box_position = "dodge2",
  violin_position = "dodge", box_stat = "boxplot",
  violin_stat = "ydensity", outlier.color = NULL,
  outlier.fill = NULL, outlier.shape = 19, outlier.size = 1.5,
  notch = FALSE, notchwidth = 0.5, add = NULL, na.rm = FALSE)
}
\arguments{
\item{df}{}

\item{grid}{= TRUE}

\item{coord_flip}{= FALSE}

\item{axis}{= FALSE}

\item{fill}{= FALSE}

\item{alpha}{= 1}

\item{fill}{= "gray30"}

\item{position}{= "stack"}

\item{theme}{= NULL}
}
\value{
plot
}
\description{
Retorna uma lista de graficos de barra, ou um grid de graficos de barra com as conjuntas de variaveis continuas e discretas de um dado dataframe
}
\examples{
conjunta_cd(iris)

}
