% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunta_scat_cc.R
\name{conjunta_scat_cc}
\alias{conjunta_scat_cc}
\title{Gera varios scatterplots com variaveis de um dataframe}
\usage{
conjunta_scat_cc(df, grid = TRUE, axis = FALSE, point_size = 0.5,
  point_shape = 1, point_color = "black", point_alpha = 0.8,
  point_position = "identity")
}
\arguments{
\item{df}{}

\item{grid}{= TRUE}

\item{axis}{= FALSE}

\item{point_size}{= 0.5}

\item{point_shape}{= 1}

\item{point_color}{= "black"}

\item{point_alpha}{= 0.8}

\item{point_position}{= "identity"}
}
\value{
plot
}
\description{
Retorna uma lista de scatterplots, ou um grid de scatterplots com as variaveis continuas de um dado dataframe
}
\examples{
conjunta_scat_cc(iris)

}
