% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_continuas.R
\name{hist_continuas}
\alias{hist_continuas}
\title{Gera varios graficos}
\usage{
hist_continuas(df, axis = FALSE, grid = TRUE, col = "black",
  fill = "gray40", stat = "bin", position = "stack",
  binwidth = NULL, bins = 10)
}
\arguments{
\item{df}{}

\item{axis}{= FALSE}

\item{grid}{= TRUE}

\item{col}{= "black"}

\item{fill}{= "gray40"}

\item{stat}{= "bin"}

\item{position}{= "stack"}

\item{binwidth}{= NULL}

\item{bins}{= 10}
}
\value{
plot
}
\description{
Retorna uma lista de graficos
}
\examples{
conjunta_bar_cd(iris)

}
