% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_continuous.R
\name{plot_continuous}
\alias{plot_continuous}
\title{Histograma}
\usage{
plot_continuous(df, grid = TRUE, plot_type = "density",
  coord_flip = FALSE, axis = FALSE, col = "black", fill = "gray40",
  alpha = 1, hist_position = "stack", box_position = "dodge2",
  violin_position = "dodge", density_position = "identity",
  hist_stat = "bin", box_stat = "boxplot", violin_stat = "ydensity",
  density_stat = "density", binwidth = NULL, bins = 10,
  outlier.color = NULL, outlier.fill = NULL, outlier.shape = 19,
  outlier.size = 1.5, notch = FALSE, notchwidth = 0.5, add = NULL,
  na.rm = FALSE)
}
\arguments{
\item{df}{}

\item{grid}{= TRUE}

\item{axis}{= FALSE}

\item{col}{= "black"}

\item{fill}{= "gray40"}

\item{binwidth}{= NULL}

\item{bins}{= 10}

\item{stat}{= "bin"}

\item{position}{= "stack"}

\item{theme}{= NULL}
}
\value{
plot
}
\description{
Retorna uma lista de graficos
}
\examples{
conjunta_bar_cd(iris)

}
