% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_discrete.R
\name{plot_discrete}
\alias{plot_discrete}
\title{Plot Discrete}
\usage{
plot_discrete(data, grid = TRUE, coord_flip = FALSE, axis = FALSE,
  factors = NULL, aes_fill = FALSE, col = "black", fill = "gray40",
  alpha = 1, position = "dodge", add = NULL)
}
\arguments{
\item{grid}{= TRUE}

\item{coord_flip}{= FALSE}

\item{axis}{= FALSE}

\item{fill}{= FALSE}

\item{df}{}

\item{factor}{= NULL}

\item{bar_fill}{= "gray40"}

\item{bar_alpha}{= 1}

\item{bar_position}{= "dodge"}

\item{theme}{= NULL}
}
\value{
plot
}
\description{
Retorna uma lista de graficos
}
\examples{
conjunta_bar_cd(iris)

}
