% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjunta_scat_cc.R
\name{conjunta_scat_cc}
\alias{conjunta_scat_cc}
\title{Gera varios scatterplots com variaveis de um dataframe}
\usage{
conjunta_scat_cc(df, grid = TRUE, axis = FALSE, size = 0.5,
  shape = 1, col = "black", alpha = 0.8, position = "identity",
  add = NULL, na.rm = FALSE)
}
\arguments{
\item{df}{}

\item{grid}{= TRUE}

\item{axis}{= FALSE}

\item{size}{= 0.5}

\item{shape}{= 1}

\item{col}{= "black"}

\item{alpha}{= 0.8}

\item{position}{= "identity"}

\item{theme}{= NULL}
}
\value{
plot
}
\description{
Retorna uma lista de scatterplots, ou um grid de scatterplots com as variaveis continuas de um dado dataframe
}
\examples{
conjunta_scat_cc(iris)

}
